Move status update email logic to email service and make async
